requirements.txt
fastapi~=0.95.0
uvicorn~=0.21.1
alembic~=1.10.3
asyncpg~=0.27.0
SQLAlchemy~=2.0.10
python-dotenv~=1.0.0
pydantic~=1.10.7
pika~=1.3.1